原来的内容可以编译运行

+++ simple_full_rt/Output_data.h
@@ -9,9 +9,9 @@ typedef struct Output_data{
     int y_position;  //目前所走到的位置y坐标
     int x_step; //x轴是否前进，前进为1，未前进则为0
     int y_step; //y轴是否前进，前进为1，未前进则为0
-    int real_cycle_time; //实际周期时间是运行平台可能实时性并不足够好，可能会有相应的波动
-    int real_intepolation_time; //由于实施周期平台不一定足够好
-    int intrp_prop_on_cycle; //差补时间占周期时间的比例
+    float real_cycle_time; //实际周期时间是运行平台可能实时性并不足够好，可能会有相应的波动
+    float real_intepolation_time; //由于实施周期平台不一定足够好
+    float intrp_prop_on_cycle; //差补时间占周期时间的比例
 }Output_data;
 
 extern struct Output_data;
